+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Initial release of FaceAISDK Flutter Plugin.
+* Support for FaceVerifyView Platform View on Android and iOS.